Fix typo in analytic package name
