--- vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/collect/core/api/analityc/AnalyticTracker.kt
@@ -1,10 +1,0 @@
-package com.verygoodsecurity.vgscheckout.collect.core.api.analityc
-
-import com.verygoodsecurity.vgscheckout.collect.core.api.analityc.event.core.Event
-
-internal interface AnalyticTracker {
-
-    var isEnabled: Boolean
-
-    fun log(event: Event)
-}
