--- vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/collect/core/api/analityc/event/CancelEvent.kt
@@ -1,10 +1,0 @@
-package com.verygoodsecurity.vgscheckout.collect.core.api.analityc.event
-
-import com.verygoodsecurity.vgscheckout.collect.core.api.analityc.event.core.Event
-
-private const val TYPE = "Cancel"
-
-internal object CancelEvent : Event(TYPE) {
-
-    override val attributes: Map<String, Any> = emptyMap()
-}
