+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/collect/core/api/analytic/AnalyticTracker.kt
@@ -1,0 +1,10 @@
+package com.verygoodsecurity.vgscheckout.collect.core.api.analytic
+
+import com.verygoodsecurity.vgscheckout.collect.core.api.analytic.event.core.Event
+
+internal interface AnalyticTracker {
+
+    var isEnabled: Boolean
+
+    fun log(event: Event)
+}
