+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/config/VGSCheckoutPaymentConfig.kt
@@ -3,7 +3,7 @@ package com.verygoodsecurity.vgscheckout.config
 import android.os.Parcel
 import android.os.Parcelable
 import com.verygoodsecurity.vgscheckout.VGSCheckoutConfigInitCallback
-import com.verygoodsecurity.vgscheckout.collect.core.api.analityc.event.JWTValidationEvent
+import com.verygoodsecurity.vgscheckout.collect.core.api.analytic.event.JWTValidationEvent
 import com.verygoodsecurity.vgscheckout.config.core.CheckoutConfig
 import com.verygoodsecurity.vgscheckout.config.networking.VGSCheckoutPaymentRouteConfig
 import com.verygoodsecurity.vgscheckout.config.ui.VGSCheckoutAddCardFormConfig
