+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/config/core/CheckoutConfig.kt
@@ -1,8 +1,8 @@
 package com.verygoodsecurity.vgscheckout.config.core
 
 import android.os.Parcelable
-import com.verygoodsecurity.vgscheckout.collect.core.api.analityc.AnalyticTracker
-import com.verygoodsecurity.vgscheckout.collect.core.api.analityc.DefaultAnalyticsTracker
+import com.verygoodsecurity.vgscheckout.collect.core.api.analytic.AnalyticTracker
+import com.verygoodsecurity.vgscheckout.collect.core.api.analytic.DefaultAnalyticsTracker
 import com.verygoodsecurity.vgscheckout.config.networking.core.CheckoutRouteConfig
 import com.verygoodsecurity.vgscheckout.config.ui.core.CheckoutFormConfig
 import com.verygoodsecurity.vgscheckout.model.VGSCheckoutEnvironment
