+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/ui/SaveCardActivity.kt
@@ -1,7 +1,7 @@
 package com.verygoodsecurity.vgscheckout.ui
 
 import android.os.Bundle
-import com.verygoodsecurity.vgscheckout.collect.core.api.analityc.event.InitEvent
+import com.verygoodsecurity.vgscheckout.collect.core.api.analytic.event.InitEvent
 import com.verygoodsecurity.vgscheckout.config.VGSCheckoutAddCardConfig
 import com.verygoodsecurity.vgscheckout.ui.core.BaseCheckoutActivity
